+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,48 @@
+package com.techcourse.service;
+
+import com.techcourse.domain.User;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+
+public class TxUserService implements UserService {
+
+    private final PlatformTransactionManager transactionManager;
+    private final UserService userService;
+
+    public TxUserService(PlatformTransactionManager transactionManager, UserService userService) {
+        this.transactionManager = transactionManager;
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void insert(User user) {
+        TransactionStatus status = transactionManager.getTransaction(new DefaultTransactionDefinition());
+
+        try {
+            userService.insert(user);
+            transactionManager.commit(status);
+        } catch (RuntimeException e) {
+            transactionManager.rollback(status);
+            throw e;
+        }
+    }
+
+    @Override
+    public void changePassword(long id, String newPassword, String createBy) {
+        TransactionStatus status = transactionManager.getTransaction(new DefaultTransactionDefinition());
+
+        try {
+            userService.changePassword(id, newPassword, createBy);
+            transactionManager.commit(status);
+        } catch (RuntimeException e) {
+            transactionManager.rollback(status);
+            throw e;
+        }
+    }
+}
