+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,46 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final PlatformTransactionManager transactionManager;
+    User findById(final long id);
 
-    public UserService(UserDao userDao, UserHistoryDao userHistoryDao, PlatformTransactionManager transactionManager) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionManager = transactionManager;
-    }
+    void insert(final User user);
 
-    public User findById(long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(long id, String newPassword, String createBy) {
-        TransactionStatus status = transactionManager.getTransaction(new DefaultTransactionDefinition());
-
-        try {
-            User user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-
-            transactionManager.commit(status);
-        } catch (RuntimeException e) {
-            transactionManager.rollback(status);
-            throw e;
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
