+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -32,8 +32,11 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var transactionManager = new DataSourceTransactionManager(DataSourceConfig.getInstance());
-        final var userService = new UserService(userDao, userHistoryDao, transactionManager);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var transactionManager = new DataSourceTransactionManager(jdbcTemplate.getDataSource());
+        final var userService = new TxUserService(transactionManager, appUserService);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -48,8 +51,11 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var transactionManager = new DataSourceTransactionManager(DataSourceConfig.getInstance());
-        final var userService = new UserService(userDao, userHistoryDao, transactionManager);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var transactionManager = new DataSourceTransactionManager(jdbcTemplate.getDataSource());
+        final var userService = new TxUserService(transactionManager, appUserService);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
